requirements.simple.txt
python-dotenv==1.0.0
pybit==5.8.0
google-generativeai==0.8.3
requests==2.31.0
flask==3.0.0
